requirements.txt
# Requirements for the project 
elevenlabs 
python-dotenv 
asyncio
ipykernel
livekit
livekit-agents[openai,cartesia,silero,turn-detector,elevenlabs]~=1.0
livekit-plugins-noise-cancellation~=0.2.4
livekit-api
livekit-plugins-deepgram~=0.6